Made alot of features working, renamed the GUI files.
You can now set settings in the settings menu, check api statuses and your ping to famous servers such as Hypixel, Mineplex, WynnCraft, MinemenClub.
Tons of errors were fixed by my guy AlphaBs (super friendly) :D
