+++ MaiksLauncher.GUI/MaiksLauncher.GUI.csproj
@@ -31,7 +31,7 @@
   <ItemGroup>
     <PackageReference Include="CefSharp.Wpf" Version="79.1.360" />
     <PackageReference Include="ChangFeng.MojangSharpCore" Version="0.0.1" />
-    <PackageReference Include="CmlLib.Core" Version="2.0.1" />
+    <PackageReference Include="CmlLib.Core" Version="2.0.2" />
     <PackageReference Include="MaterialDesignThemes" Version="3.1.3" />
     <PackageReference Include="Microsoft.Xaml.Behaviors.Wpf" Version="1.1.19" />
     <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
