+++ MaiksLauncher.GUI/MainWindowOld.xaml
@@ -1,0 +1,58 @@
+﻿<Window x:Class="MaiksLauncher.GUI.MainWindow"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:wpf="clr-namespace:CefSharp.Wpf;assembly=CefSharp.Wpf"
+
+        mc:Ignorable="d"
+        Title="MaiksLauncher" Height="436" Width="800" ResizeMode="NoResize" Icon="/Photos/Steve.png" Loaded="winactive">
+    <Grid Margin="0,0,0,-16">
+        <Grid.RowDefinitions>
+            <RowDefinition Height="282*" />
+            <RowDefinition Height="137*" />
+        </Grid.RowDefinitions>
+        <Grid.ColumnDefinitions>
+            <ColumnDefinition Width="7*" />
+            <ColumnDefinition Width="153*" />
+        </Grid.ColumnDefinitions>
+        <TextBox x:Name="Email" HorizontalAlignment="Left" Margin="42,62,0,0" TextWrapping="NoWrap" VerticalAlignment="Top" Width="135"   Grid.Column="1" Grid.Row="1" AllowDrop="True" />
+        <Label Content="Email:" HorizontalAlignment="Left" Margin="10,58,0,0" VerticalAlignment="Top" RenderTransformOrigin="0.091,1.711" Grid.ColumnSpan="2" Grid.Row="1" />
+        <PasswordBox Height="18"  Margin="42,87,588,0" x:Name="Password" VerticalAlignment="Top" PasswordChar="*" Grid.Column="1" Grid.Row="1" />
+        <Label Content="Password:" HorizontalAlignment="Left" Margin="10,80,0,0" VerticalAlignment="Top" Grid.ColumnSpan="2" Grid.Row="1" />
+        <CheckBox Content="Save password/email" HorizontalAlignment="Left" Margin="176,59,0,0" VerticalAlignment="Top" RenderTransformOrigin="0.197,0.637" Grid.Column="1" Grid.Row="1" Visibility="Hidden" />
+        <Button Content="Launch" HorizontalAlignment="Left" Margin="201,79,0,0" VerticalAlignment="Top" Width="97" Grid.Column="1" Grid.Row="1" />
+
+        <Button x:Name="OpenLog" Content="New look" Grid.Column="1" Margin="556,71,44,40" Grid.Row="1" Click="OpenLog_Click" Background="{x:Null}" BorderBrush="{x:Null}" Foreground="Black" />
+        <Label x:Name="LoginSuccess" Content="" Grid.Column="1" HorizontalAlignment="Left" Margin="284,79,0,0" Grid.Row="1" VerticalAlignment="Top" Foreground="Lime" />
+        <Label x:Name="LoginUnsuccess" Content="" Grid.Column="1" HorizontalAlignment="Left" Margin="284,79,0,0" Grid.Row="1" VerticalAlignment="Top" Foreground="Red" />
+        <TabControl Grid.ColumnSpan="2" Margin="0,0,0,97" Grid.RowSpan="2">
+            <TabItem Header="Versions" BorderBrush="Black">
+                <Grid Margin="0,0,-38,0" Height="297">
+                    <Grid.ColumnDefinitions>
+                        <ColumnDefinition Width="29*" />
+                        <ColumnDefinition Width="85*"/>
+                        <ColumnDefinition Width="61*" />
+                        <ColumnDefinition Width="241*"/>
+                    </Grid.ColumnDefinitions>
+
+                </Grid>
+            </TabItem>
+            <TabItem Header="TabItem">
+                <Grid>
+                    <TextBox x:Name="CustomMCDir" HorizontalAlignment="Left" Margin="130,39,0,0" TextWrapping="Wrap" VerticalAlignment="Top" Width="637" />
+                    <Label Content="Minecraft Directory:" Margin="10,33,664,231" />
+                    <TextBox HorizontalAlignment="Left" Margin="130,75,0,0" Text="TextBox" TextWrapping="Wrap" VerticalAlignment="Top" Width="637" />
+                    <TextBox HorizontalAlignment="Left" Margin="130,113,0,0" Text="TextBox" TextWrapping="Wrap" VerticalAlignment="Top" Width="637" />
+                    <TextBox HorizontalAlignment="Left" Margin="130,147,0,0" Text="TextBox" TextWrapping="Wrap" VerticalAlignment="Top" Width="637" />
+                    <TextBox HorizontalAlignment="Left" Margin="130,185,0,0" Text="TextBox" TextWrapping="Wrap" VerticalAlignment="Top" Width="637" />
+                    <Label Content="Custom Arguments:" HorizontalAlignment="Left" Margin="10,71,0,0" VerticalAlignment="Top" />
+                    <Label Content="Log save folder:" HorizontalAlignment="Left" Margin="10,109,0,0" VerticalAlignment="Top" />
+                    <Label Content="&lt;empty&gt;" HorizontalAlignment="Left" Margin="10,143,0,0" VerticalAlignment="Top" />
+                    <Label Content="&lt;empty&gt;" HorizontalAlignment="Left" Margin="10,176,0,0" VerticalAlignment="Top" />
+                </Grid>
+            </TabItem>
+        </TabControl>
+        <ComboBox x:Name="LauncherProfiles" Grid.Column="1" HorizontalAlignment="Left" Margin="201,51,0,0" Grid.Row="1" VerticalAlignment="Top" Width="97" Height="26"/>
+    </Grid>
+</Window>
